+++ src/editor/dataset/constant/Element.ts
@@ -13,6 +13,8 @@ export const EDITOR_ELEMENT_STYLE_ATTR: Array<keyof IElement> = [
   'strikeout',
 ]
 
+export const EDITOR_ROW_ATTR: Array<keyof IElement> = ['rowFlex', 'rowMargin']
+
 export const EDITOR_ELEMENT_COPY_ATTR: Array<keyof IElement> = [
   'type',
   'font',
@@ -29,6 +31,8 @@ export const EDITOR_ELEMENT_COPY_ATTR: Array<keyof IElement> = [
   'dateId',
   'dateFormat',
   'groupIds',
+  'rowFlex',
+  'rowMargin'
 ]
 
 export const EDITOR_ELEMENT_ZIP_ATTR: Array<keyof IElement> = [