+++ src/editor/index.ts
@@ -70,6 +70,9 @@ import { IRange } from './interface/Range'
 import { deepClone, splitText } from './utils'
 import { IZoneOption } from './interface/Zone'
 import { defaultZoneOption } from './dataset/constant/Zone'
+import { IBackgroundOption } from './interface/Background'
+import { defaultBackground } from './dataset/constant/Background'
+import { BackgroundRepeat, BackgroundSize } from './dataset/enum/Background'
 
 export default class Editor {
   public command: Command
@@ -137,10 +140,15 @@ export default class Editor {
       ...defaultZoneOption,
       ...options.zone,
     }
+    const backgroundOptions: Required<IBackgroundOption> = {
+      ...defaultBackground,
+      ...options.background
+    }
 
     const editorOptions: DeepRequired<IEditorOption> = {
       mode: EditorMode.EDIT,
       defaultType: 'TEXT',
+      defaultColor: '#000000',
       defaultFont: 'Microsoft YaHei',
       defaultSize: 16,
       minSize: 5,
@@ -152,7 +160,6 @@ export default class Editor {
       height: 1123,
       scale: 1,
       pageGap: 20,
-      backgroundColor: '#FFFFFF',
       underlineColor: '#000000',
       strikeoutColor: '#FF0000',
       rangeAlpha: 0.6,
@@ -195,6 +202,7 @@ export default class Editor {
       group: groupOptions,
       pageBreak: pageBreakOptions,
       zone: zoneOptions,
+      background: backgroundOptions
     }
     // 数据处理
     data = deepClone(data)
@@ -289,6 +297,8 @@ export {
   ListStyle,
   WordBreak,
   ControlIndentation,
+  BackgroundRepeat,
+  BackgroundSize
 }
 
 // 对外类型